Move some suggestions from error_reporting to error_reporting::suggest
